fix rpath of bianry to correct value.
The old value of rpath is an absolute path on build host, which is
not exist on target machine. Change it to relative path.
After this commit, LD_LIBRARY_PATH could be removed. Problem is that
we still have a lot of extra dependency for enterprise building, so
LD_LIBRARY_PATH is still kept.